requirements.txt
fastapi 
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary 
aiosqlite 
asyncpg 
grpcio 
grpcio-tools
asyncio
types-protobuf
